pyproject.toml
[project]
name = "geopol-forecaster"
version = "0.1.0"
description = "Two-stage geopolitical forecaster: snowglobe actor simulation + llm-council lens panel."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "typer>=0.12",
    "python-dotenv>=1.0",
    "feedparser>=6.0",
    "tavily-python>=0.5",
    "rich>=13.7",
]

# Note: llm-snowglobe inspired the Stage A pattern but is NOT installed by
# default — its optional stack (torch, transformers, llama-cpp-python) is
# heavy and unnecessary for our direct-OpenRouter implementation. Install it
# manually (`uv pip install llm-snowglobe`) only if swapping in the real
# Control/Player classes.

[project.scripts]
geopol = "geopol.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["geopol"]
